+++ main.py
@@ -24,7 +24,8 @@ running = True
 while running:
     screen.fill((255, 255, 255))
 
-    for event in pygame.event.get():
+    events = pygame.event.get()
+    for event in events:
         if event.type == pygame.QUIT:
             running = False
         if event.type == pygame.KEYDOWN and event.key == pygame.K_ESCAPE:
@@ -33,14 +34,11 @@ while running:
             pygame.mouse.set_visible(0)
             sprite.rect.x = pygame.mouse.get_pos()[0]
             sprite.rect.y = pygame.mouse.get_pos()[1]
-            if event.type == pygame.MOUSEBUTTONDOWN:
-                if event.button == 1:
-                    level.fire(event.pos, level.gun.rect[:2])
         # TODO На Esc можно сделать смену режима окна (полный/неполный экран), я не придумал как чтобы все работало
         # elif event.type == pygame.KEYDOWN and event.key == pygame.K_ESCAPE:
         #     pygame.display.set_mode(size, flags=pygame.RESIZABLE)
 
-    level.update()
+    level.update(events)
 
     all_sprites.draw(screen)
     pygame.display.flip()
